int64-

+++ product_service.go
@@ -1,33 +1,31 @@
 package binance
 
-
 import (
-"context"
-"encoding/json"
+	"context"
+	"encoding/json"
 )
 
 // ListBookTickersService list all book tickers
 type ListProductService struct {
-    c *Client
+	c *Client
 }
 
 // Do send request
 func (s *ListProductService) Do(ctx context.Context, opts ...RequestOption) (p ProductList, err error) {
-    r := &request{
-        method:   "GET",
-        endpoint: "/exchange/public/product",
-    }
-    data, err := s.c.callAPI(ctx, r, opts...)
-    if err != nil {
-        return
-    }
-    err = json.Unmarshal(data, &p)
-    if err != nil {
-        return
-    }
-    return
+	r := &request{
+		method:   "GET",
+		endpoint: "/exchange/public/product",
+	}
+	data, err := s.c.callAPI(ctx, r, opts...)
+	if err != nil {
+		return
+	}
+	err = json.Unmarshal(data, &p)
+	if err != nil {
+		return
+	}
+	return
 }
-
 
 /*
 {
@@ -59,29 +57,29 @@ func (s *ListProductService) Do(ctx context.Context, opts ...RequestOption) (p P
 */
 // PriceChangeStats define price change stats
 type ProductList struct {
-    Data []Product `json:"data"`
+	Data []Product `json:"data"`
 }
 
 type Product struct {
-    Symbol           string  `json:"symbol"`
-    TradedMoney      float64 `json:"tradedMoney"`
-    Active           bool    `json:"active"`
-    MinTrade         float64  `json:"minTrade,string"`
-    BaseAsset        string  `json:"baseAsset"`
-    ActiveSell       float64 `json:"activeSell"`
-    WithdrawFee      float64 `json:"withdrawFee,string"`
-    TickSize         float64 `json:"tickSize,string"`
-    PrevClose        float64 `json:"prevClose"`
-    ActiveBuy        int64   `json:"activeBuy"`
-    Volume           float64 `json:"volume,string"`
-    High             float64 `json:"high,string"`
-    LastAggTradeId   int64   `json:"lastAggTradeId"`
-    DecimalPlaces    int64   `json:"decimalPlaces"`
-    QuoteAssetUnit   string  `json:"quoteAssetUnit"`
-    MatchingUnitType string  `json:"matchingUnitType"`
-    Close            float64 `json:"close,string"`
-    QuoteAsset       string  `json:"quoteAsset"`
-    Open             float64 `json:"open,string"`
-    Status           string  `json:"status"`
-    MinQty           float64 `json:"minQty,string"`
+	Symbol           string  `json:"symbol"`
+	TradedMoney      float64 `json:"tradedMoney"`
+	Active           bool    `json:"active"`
+	MinTrade         float64 `json:"minTrade,string"`
+	BaseAsset        string  `json:"baseAsset"`
+	ActiveSell       float64 `json:"activeSell"`
+	WithdrawFee      float64 `json:"withdrawFee,string"`
+	TickSize         float64 `json:"tickSize,string"`
+	PrevClose        float64 `json:"prevClose"`
+	ActiveBuy        float64 `json:"activeBuy"`
+	Volume           float64 `json:"volume,string"`
+	High             float64 `json:"high,string"`
+	LastAggTradeId   int64   `json:"lastAggTradeId"`
+	DecimalPlaces    int64   `json:"decimalPlaces"`
+	QuoteAssetUnit   string  `json:"quoteAssetUnit"`
+	MatchingUnitType string  `json:"matchingUnitType"`
+	Close            float64 `json:"close,string"`
+	QuoteAsset       string  `json:"quoteAsset"`
+	Open             float64 `json:"open,string"`
+	Status           string  `json:"status"`
+	MinQty           float64 `json:"minQty,string"`
 }
